safari-600.7-branch/LayoutTests/http/tests/w3c/webperf/submission/Google/resource-timing/html/test_resource_redirects.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing redirect names</title>
<link rel="author" title="Google" href="http://www.google.com/" />
<link rel="help" href="http://www.w3.org/TR/resource-timing/#performanceresourcetiming"/>
<script src="/w3c/resources/testharness.js"></script>
<script src="/w3c/resources/testharnessreport.js"></script>
<script src="/w3c/webperf/resources/webperftestharness.js"></script>
<script src="/w3c/webperf/resources/webperftestharnessextension.js"></script>
<script>
var iframe;
function setup_iframe() {
    var iframe_content =
        '<link rel="stylesheet" href="/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=css"></link>' +
        '<img src="/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=image"></img>' +
        '<iframe src="/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=iframe"></iframe>' +
        '<script src="/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=script"></scr' + 'ipt>' +
        '<script>' +
        'var xhr = new XMLHttpRequest;' +
        'xhr.open("GET", "/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=iframe%26id=xhr", false);' +
        'xhr.send();' +
        '</scr' + 'ipt>';
    iframe = document.getElementById('frameContext');
    iframe.contentWindow.document.write(iframe_content);
}
function onload_test() {
    var context = new PerformanceContext(iframe.contentWindow.performance);
    var entries = context.getEntries();

    var expected_entries = {
        '/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=css': 'link',
        '/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=image': 'img',
        '/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=iframe': 'iframe',
        '/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=script': 'script',
        '/w3c/webperf/resources/redirect.php?location=/w3c/webperf/resources/generate_resource.php%3Ftype=iframe%26id=xhr': 'xmlhttprequest',
    };

    test_resource_entries(entries, expected_entries);
}
window.setup_iframe = setup_iframe;
</script>
</head>
<body>
<h1>Description</h1>
<p>This test validates that redirects do not alter the URL.</p>
<div id="log"></div>
<iframe id="frameContext" onload="onload_test();" src="/w3c/webperf/resources/inject_resource_test.html"></iframe>
</body>
</html>
